backend/pyproject.toml
[project]
name = "pramana-ai"
version = "1.0.0"
description = "Pramana.ai - Master biomedical evidence. Without the silos."
authors = [
    { name = "Team GitGonewild" },
]
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11,<4.0"
keywords = [
    "ai",
    "biomedical",
    "research",
    "langgraph",
    "langchain",
    "groq",
    "agents",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langgraph>=0.2.6",
    "langchain>=0.3.19",
    "langchain-groq",
    "langchain-community",
    "google-search-results",
    "python-dotenv>=1.0.1",
    "langgraph-sdk>=0.1.57",
    "langgraph-cli",
    "langgraph-api",
    "fastapi",
    "langchain-mcp-adapters>=0.1.7",
    # Auth dependencies
    "sqlalchemy>=2.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "pydantic[email]>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/N1KH1LT0X1N/Pharma-Agent"
Repository = "https://github.com/N1KH1LT0X1N/Pharma-Agent"
Documentation = "https://github.com/N1KH1LT0X1N/Pharma-Agent#readme"

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",  # Allow typing_extensions imports
    "D417",   # Don't require documentation for every function parameter
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.1.71",
    "pytest>=8.3.5",
    "ruff>=0.11.12",
]
